requirements.txt
python-dotenv
tweepy
textblob
